ImportKeys: Move creation date in the top part of the card
