+++ OpenKeychain/src/main/res/layout/import_keys_list_item.xml
@@ -28,39 +28,39 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical"
                 android:paddingBottom="16dp"
+                android:paddingLeft="16dp"
+                android:paddingRight="16dp"
                 android:paddingTop="24dp">
 
                 <LinearLayout
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:orientation="vertical"
-                    android:paddingLeft="16dp"
-                    android:paddingRight="16dp">
+                    android:orientation="vertical">
 
-                    <LinearLayout
-                        android:layout_width="match_parent"
+                    <TextView
+                        android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:orientation="vertical">
+                        android:textAppearance="?android:attr/textAppearanceMedium"
+                        app:keyRevokedOrExpired="@{entry.revokedOrExpired}"
+                        app:keySecret="@{entry.secretKey}"
+                        app:keyUserId="@{entry.primaryUserId.name}"
+                        app:query="@{entry.query}" />
 
-                        <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:textAppearance="?android:attr/textAppearanceMedium"
-                            app:keyRevokedOrExpired="@{entry.revokedOrExpired}"
-                            app:keySecret="@{entry.secretKey}"
-                            app:keyUserId="@{entry.primaryUserId.name}"
-                            app:query="@{entry.query}" />
+                    <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="4dp"
+                        android:textAppearance="?android:attr/textAppearanceSmall"
+                        app:keyRevokedOrExpired="@{entry.revokedOrExpired}"
+                        app:keyUserEmail="@{entry.primaryUserId.email}"
+                        app:query="@{entry.query}" />
 
-                        <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_marginTop="4dp"
-                            android:textAppearance="?android:attr/textAppearanceSmall"
-                            app:keyRevokedOrExpired="@{entry.revokedOrExpired}"
-                            app:keyUserEmail="@{entry.primaryUserId.email}"
-                            app:query="@{entry.query}" />
-
-                    </LinearLayout>
+                    <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:textAppearance="?android:attr/textAppearanceSmall"
+                        app:keyCreation="@{entry.date}"
+                        app:keyRevokedOrExpired="@{entry.revokedOrExpired}" />
 
                 </LinearLayout>
