+++ OpenKeychain/src/main/res/layout/import_keys_list_item_extra.xml
@@ -77,27 +77,6 @@
 
                     </LinearLayout>
 
-                    <LinearLayout
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content">
-
-                        <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:text="@string/label_creation_colon"
-                            android:textAppearance="?android:attr/textAppearanceSmall"
-                            android:textColor="?attr/colorText" />
-
-                        <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_marginLeft="8dp"
-                            android:textAppearance="?android:attr/textAppearanceSmall"
-                            android:textColor="?attr/colorText"
-                            app:keyCreation="@{entry.date}" />
-
-                    </LinearLayout>
-
                 </LinearLayout>
 
                 <ImageView
